+++ src/repositories/portable_repo.rs
@@ -50,9 +50,6 @@ pub enum PortableRepoError {
     #[error("The destination does already exist and is not an empty directory")]
     DestinationNotEmpty,
 
-    #[error("The given package name is empty")]
-    EmptyPackageName,
-
     #[error("Cannot fetch package metadata from repository")]
     RepositoryError(#[from] RepositoryError),
 
@@ -114,18 +111,7 @@ impl<'a> PortableRepoCreator<'a> {
                 self.repository_manager.read_package_and_version(&package_id.clone().into(), &self.target)?;
 
             if !self.exclude_prebuilds {
-                // Check if the package has a prebuild
-                let prebuild_meta = self.repository_manager.get_prebuild_meta(
-                    &repository_id,
-                    &package_id,
-                    package_version.get_revision_count(),
-                    &self.target.architecture.to_string(),
-                );
-                let found_prebuild = match prebuild_meta {
-                    Ok(_) => true,
-                    Err(RepositoryError::PrebuildNotFound { .. }) => false,
-                    Err(e) => return Err(e.into()),
-                };
+                let found_prebuild = self.is_prebuild_available(&repository_id, &package_id, package_version.get_revision_count())?;
 
                 // Check if prebuild is downloadable, or if package is installed
                 if !found_prebuild && (self.target != Target::current() || self.register.get_package_version(&package_id).is_none()) {
@@ -169,6 +155,24 @@ impl<'a> PortableRepoCreator<'a> {
         self.write_metadata(index_meta, destination.join("index.toml"), true)?;
 
         Ok(())
+    }
+
+    /// Checks if a prebuild is available for the given package and revision in the given repository.
+    /// Returns `true` if a prebuild is available, false otherwise.
+    fn is_prebuild_available(&self, repository_id: &str, package_id: &PackageId, revision: u64) -> Result<bool> {
+        // Get id of the prebuild
+        let prebuilds_list = self.repository_manager.read_prebuilds_list(repository_id, &package_id.name, &package_id.version)?;
+        let Some((prebuild_id, _)) = prebuilds_list.get_best_prebuild(&Target::current()) else {
+            return Ok(false);
+        };
+
+        // Check if the package has a prebuild
+        let prebuild_meta = self.repository_manager.get_prebuild_meta(repository_id, package_id, revision, prebuild_id);
+        match prebuild_meta {
+            Ok(_) => Ok(true),
+            Err(RepositoryError::PrebuildNotFound { .. }) => Ok(false),
+            Err(e) => Err(e.into()),
+        }
     }
 
     /// Creates a set of all packages that are in the dependency trees of the included packages.
@@ -258,6 +262,11 @@ impl<'a> PortableRepoCreator<'a> {
         let targets_path = package_path.join(package_id.version.to_string()).join("targets.toml");
         self.write_metadata(version_meta, targets_path, false)?;
 
+        // Download prebuilds.toml
+        let prebuilds_list = self.repository_manager.read_prebuilds_list(repository_id, &package_id.name, &package_id.version)?;
+        let prebuilds_path = package_path.join(package_id.version.to_string()).join("prebuilds.toml");
+        self.write_metadata(prebuilds_list, prebuilds_path, false)?;
+
         let target_bounds = package_version.get_best_target(&self.target)?;
 
         // Download scripts
@@ -292,30 +301,35 @@ impl<'a> PortableRepoCreator<'a> {
         package_version: &PackageVersionMeta,
         destination: &Path,
     ) -> Result<()> {
-        let prefix = package_id.name.chars().next().ok_or(PortableRepoError::EmptyPackageName)?.to_string();
-        let target = self.target.architecture.to_string();
-        let destination = destination.join("prebuilds").join(&target).join(&prefix).join(&package_id.name);
+        let prefix = package_id.name.get_prefix().to_string();
+        let prebuilds_dir = destination.join("prebuilds");
+        let destination = prebuilds_dir.join("packages").join(prefix).join(&package_id.name).join(package_id.version.to_string());
         fs::create_dir_all(&destination).err_with_path("create dirs", &destination)?;
+
+        // Get id of the prebuild
+        let prebuilds_list = self.repository_manager.read_prebuilds_list(repository_id, &package_id.name, &package_id.version)?;
+        let (prebuild_id, _) = prebuilds_list.get_best_prebuild(&Target::current()).ok_or(PortableRepoError::PrebuildNotFound {
+            package_id: package_id.clone(),
+        })?;
 
         // Get checksum
         let revision = package_version.get_revision_count();
-        let prebuild_id = self.target.architecture.to_string();
-        let prebuild_meta = match self.repository_manager.get_prebuild_meta(repository_id, package_id, revision, &prebuild_id) {
+        let prebuild_meta = match self.repository_manager.get_prebuild_meta(repository_id, package_id, revision, prebuild_id) {
             Ok(prebuild_meta) => prebuild_meta,
             // Only try to package locally if the current target is the target we generate a portable repo for
             Err(RepositoryError::PrebuildNotFound { .. }) if self.target == Target::current() => {
-                packager::package(self.config, package_id, &destination, revision)?;
+                packager::package(self.config, package_id, &destination, revision, prebuild_id)?;
                 return Ok(());
             },
             Err(e) => return Err(e.into()),
         };
 
-        let (_, prebuild) = self.repository_manager.read_prebuild(repository_id, package_id, revision, &prebuild_id)?;
+        let (_, prebuild) = self.repository_manager.read_prebuild(repository_id, package_id, revision, prebuild_id)?;
 
         // Write to file
-        let prebuild_name = format!("{package_id}-{revision}-{target}.tar.gz");
+        let prebuild_name = format!("{package_id}-{revision}-{prebuild_id}.tar.gz");
         let prebuild_path = destination.join(prebuild_name);
-        let prebuild_meta_name = format!("{package_id}-{revision}-{target}.toml");
+        let prebuild_meta_name = format!("{package_id}-{revision}-{prebuild_id}.toml");
         let prebuild_meta_path = destination.join(prebuild_meta_name);
         fs::write(&prebuild_path, &prebuild).err_with_path("write", prebuild_path)?;
         self.write_metadata(prebuild_meta, prebuild_meta_path, false)?;