+++ src/repositories/types/mod.rs
@@ -19,6 +19,9 @@ pub use self::package_target::PackageTarget;
 pub use self::package_version::PackageVersionMeta;
 
 pub use self::prebuilds::PrebuildFileMeta;
+#[expect(unused_imports)]
+pub use self::prebuilds::PrebuildMeta;
+pub use self::prebuilds::PrebuildsList;
 
 pub use self::checksum::Checksum;
 pub use self::common::Script;